+++ plugins/test/src/lib.rs
@@ -1,3 +1,5 @@
-#[no_mangle]
-/// This function is called by the Vienna game engine on every update.
-pub extern "C" fn _run() {}
+vienna::load!();
+
+fn run() -> Result<()> {
+    Ok(())
+}